requirements.txt
PyQt6>=6.5.0
mss>=9.0.0
Pillow>=10.0.0
cryptography>=41.0.0
